+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Release Notes for Campaign
 
+## 1.24.5 - Unreleased
+### Fixed
+- Fixed the status filter for campaigns, mailing lists and contacts selectable in sendouts ([#419](https://github.com/putyourlightson/craft-campaign/issues/419)).
+
 ## 1.24.4 - 2023-09-20
 ### Fixed
 - Fixed a bug in which the cached sendout element view was not invalidated after a sendout completed ([#416](https://github.com/putyourlightson/craft-campaign/issues/416)).